requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
supabase>=2.0.0
stripe>=7.0.0
httpx>=0.25.0
pydantic>=2.4.0
weasyprint>=60.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
email-validator>=2.1.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.0.0